Merge branch 'master' into task_6938673

+++ src/backend/ci/ext/tencent/process/biz-process-tencent/src/main/kotlin/com/tencent/devops/process/engine/atom/task/QualityGateOutTaskAtom.kt
@@ -139,7 +139,7 @@ class QualityGateOutTaskAtom @Autowired constructor(
             logger.error("[$buildId]|QUALITY_OUT|taskId=$elementId|Fail to find quality gate intercept element")
             throw BuildTaskException(
                 errorType = ErrorType.USER,
-                errorCode = ERROR_BUILD_TASK_QUALITY_OUT_INTERCEPT,
+                errorCode = ERROR_BUILD_TASK_QUALITY_OUT_INTERCEPT.toInt(),
                 errorMsg = "Fail to find quality gate intercept element"
             )
         }
