+++ src/backend/ci/ext/tencent/process/biz-process-tencent/src/main/kotlin/com/tencent/devops/process/engine/atom/task/ZhiyunUpdateAsyncEXTaskAtom.kt
@@ -124,7 +124,7 @@ class ZhiyunUpdateAsyncEXTaskAtom @Autowired constructor(
                 LogUtils.addRedLine(rabbitTemplate, task.buildId, "织云异步升级失败,织云任务执行超时", task.taskId, task.containerHashId, task.executeCount ?: 1)
                 throw BuildTaskException(
                     errorType = ErrorType.SYSTEM,
-                    errorCode = ERROR_BUILD_TASK_ZHIYUN_UPGRADE_FAIL,
+                    errorCode = ERROR_BUILD_TASK_ZHIYUN_UPGRADE_FAIL.toInt(),
                     errorMsg = "织云异步升级失败,织云任务执行超时"
                 )
             }
@@ -166,7 +166,7 @@ class ZhiyunUpdateAsyncEXTaskAtom @Autowired constructor(
                 LogUtils.addRedLine(rabbitTemplate, task.buildId, "织云异步升级失败,织云返回错误信息：$msg", task.taskId, task.containerHashId, task.executeCount ?: 1)
                 throw BuildTaskException(
                     errorType = ErrorType.SYSTEM,
-                    errorCode = ERROR_BUILD_TASK_ZHIYUN_UPGRADE_FAIL,
+                    errorCode = ERROR_BUILD_TASK_ZHIYUN_UPGRADE_FAIL.toInt(),
                     errorMsg = "织云异步升级失败,织云返回错误信息：$msg"
                 )
             } else {
@@ -181,7 +181,7 @@ class ZhiyunUpdateAsyncEXTaskAtom @Autowired constructor(
                         LogUtils.addRedLine(rabbitTemplate, task.buildId, "织云异步升级失败,织云返回错误信息: errmsg：$errmsg, lastErrmsg: $lastErrmsg", task.taskId, task.containerHashId, task.executeCount ?: 1)
                         throw BuildTaskException(
                             errorType = ErrorType.SYSTEM,
-                            errorCode = ERROR_BUILD_TASK_ZHIYUN_UPGRADE_FAIL,
+                            errorCode = ERROR_BUILD_TASK_ZHIYUN_UPGRADE_FAIL.toInt(),
                             errorMsg = "织云异步升级失败,织云返回错误信息: errmsg：$errmsg, lastErrmsg: $lastErrmsg"
                         )
                     } else if (0 == status) {
@@ -214,7 +214,7 @@ class ZhiyunUpdateAsyncEXTaskAtom @Autowired constructor(
                 LogUtils.addRedLine(rabbitTemplate, task.buildId, "织云异步升级失败,织云返回错误信息：$msg", task.taskId, task.containerHashId, task.executeCount ?: 1)
                 throw BuildTaskException(
                     errorType = ErrorType.SYSTEM,
-                    errorCode = ERROR_BUILD_TASK_ZHIYUN_UPGRADE_FAIL,
+                    errorCode = ERROR_BUILD_TASK_ZHIYUN_UPGRADE_FAIL.toInt(),
                     errorMsg = "织云异步升级失败,织云返回错误信息：$msg"
                 )
             } else {
